Better mobile support, loading prop for buttons, fixed data-state-open CSS stuff
Added better CSS support for mobiles down to S size, including making the navbar auto adjust to screen size.

The loading prop for the shadcn UI buttons allows for feedback for when forms are in the process of submitting and are yet to have a response. It uses the new LoadingSpinner.tsx that is also used when loading a post.

A very nasty CSS override was going on when dropdown menus were opened. This has been remedied for now by taking the classes to the parent html element instead of the body, so they aren't overridden.