New email confirmation flow, added simple meta and post & module fetch routes.
- New Email Confirmation Flow
Email links to a capture page that requires user interaction to confirm, ensuring prefetching doesn't consume any token.

- AuthWatcher.tsx
Simple Client Component that allows for reactivity based on auth activities, used in the confirm page to redirect when a sign in event is detected.

- Metadata
Really simple metadata added, and a test of dynamic images in the post route.

- Post & Module routes
Fetch routes for singular posts and modules added to service the dynamic image creation.

- use-media-query.ts
Added from the shadcn hooks to allow for comment replies open in a dialog box if the screen is too narrow.